+++ .gitignore
@@ -4,3 +4,6 @@ build
 __pycache__
 *~
 sdkconfig.old
+.config
+*.bin
+*.elf
