+++ bsp/common/bl808/include/bl808_board.h
@@ -1,0 +1,7 @@
+#ifndef BL808_BOARD_H
+#define BL808_BOARD_H
+#include "sdkconfig.h"
+
+void board_common_init(void);
+
+#endif
