+++ bsp/ox64/Kconfig
@@ -1,20 +1,11 @@
-config JTAG_DEBUG
-    bool "JTAG debug support"
-    help
-      This option enables JTAG debug support for the test kernel.
+config BOARD_OX64
+    bool
+    default y if BFLB_BOARD = "ox64"
 
-menu "JTAG Configuration"
-    visible if JTAG_DEBUG
-    choice
-        prompt "JTAG Pinmux"
-        default JTAG_M0
-        help
-            Select the core to attach the JTAG Debug Socket to.
-        config JTAG_M0
-            bool "Debug M0 Core"
-        config JTAG_D0
-            bool "Debug D0 Core"
-        config JTAG_LP
-            bool "Debug LP Core"
-    endchoice
-endmenu
+config BOARD_HAS_SDCARD
+    bool
+    default y if BFLB_BOARD = "ox64"
+
+comment "ox64 Options"
+
+
