+++ bsp/ox64/board.c
@@ -1,579 +1,14 @@
-#include "bflb_uart.h"
-#include "bflb_gpio.h"
-#include "bflb_clock.h"
-#include "bflb_rtc.h"
-#include "bflb_flash.h"
-#include "bl808_glb.h"
-#include "bl808_psram_uhs.h"
-#include "bl808_tzc_sec.h"
-#include "bl808_ef_cfg.h"
-#include "bl808_uhs_phy.h"
-#include "board.h"
+#include <bflb_gpio.h>
+#include <bflb_clock.h>
+#include "bl808_board.h"
+#include "log.h"
 
-#include "mem.h"
 
-#ifdef CONFIG_BSP_SDH_SDCARD
-#include "sdh_sdcard.h"
-#endif
 
-extern void log_start(void);
 
-extern uint32_t __HeapBase;
-extern uint32_t __HeapLimit;
 
-static struct bflb_device_s *uart0;
+void board_init() {
+    board_common_init();
 
-#if (defined(CONFIG_LUA) || defined(CONFIG_BFLOG) || defined(CONFIG_FATFS))
-static struct bflb_device_s *rtc;
-#endif
-
-#if defined(CPU_M0)
-static void system_clock_init(void)
-{
-    /* wifipll/audiopll */
-    GLB_Power_On_XTAL_And_PLL_CLK(GLB_XTAL_40M, GLB_PLL_WIFIPLL |
-                                                    GLB_PLL_CPUPLL |
-                                                    GLB_PLL_UHSPLL |
-                                                    GLB_PLL_MIPIPLL);
-
-    GLB_Set_MCU_System_CLK(GLB_MCU_SYS_CLK_WIFIPLL_320M);
-    GLB_Set_DSP_System_CLK(GLB_DSP_SYS_CLK_CPUPLL_400M);
-    GLB_Config_CPU_PLL(GLB_XTAL_40M, cpuPllCfg_480M);
-
-    CPU_Set_MTimer_CLK(ENABLE, CPU_Get_MTimer_Source_Clock() / 1000 / 1000 - 1);
 }
 
-static void peripheral_clock_init(void)
-{
-    PERIPHERAL_CLOCK_ADC_DAC_ENABLE();
-    PERIPHERAL_CLOCK_SEC_ENABLE();
-    PERIPHERAL_CLOCK_DMA0_ENABLE();
-    PERIPHERAL_CLOCK_UART0_ENABLE();
-    PERIPHERAL_CLOCK_UART1_ENABLE();
-    PERIPHERAL_CLOCK_SPI0_1_ENABLE();
-    PERIPHERAL_CLOCK_I2C0_ENABLE();
-    PERIPHERAL_CLOCK_PWM0_ENABLE();
-    PERIPHERAL_CLOCK_TIMER0_1_WDG_ENABLE();
-    PERIPHERAL_CLOCK_IR_ENABLE();
-    PERIPHERAL_CLOCK_I2S_ENABLE();
-    PERIPHERAL_CLOCK_USB_ENABLE();
-    PERIPHERAL_CLOCK_CAN_UART2_ENABLE();
-
-    GLB_Set_ADC_CLK(ENABLE, GLB_ADC_CLK_XCLK, 4);
-    GLB_Set_UART_CLK(ENABLE, HBN_UART_CLK_XCLK, 0);
-    GLB_Set_DSP_UART0_CLK(ENABLE, GLB_DSP_UART_CLK_DSP_XCLK, 0);
-    GLB_Set_SPI_CLK(ENABLE, GLB_SPI_CLK_MCU_MUXPLL_160M, 0);
-    GLB_Set_I2C_CLK(ENABLE, GLB_I2C_CLK_XCLK, 0);
-    GLB_Set_IR_CLK(ENABLE, GLB_IR_CLK_SRC_XCLK, 19);
-    GLB_Set_ADC_CLK(ENABLE, GLB_ADC_CLK_XCLK, 1);
-    GLB_Set_DIG_CLK_Sel(GLB_DIG_CLK_XCLK);
-    GLB_Set_DIG_512K_CLK(ENABLE, ENABLE, 0x4E);
-    GLB_Set_PWM1_IO_Sel(GLB_PWM1_IO_DIFF_END);
-    GLB_Set_CAM_CLK(ENABLE, GLB_CAM_CLK_WIFIPLL_96M, 3);
-
-    GLB_Set_PKA_CLK_Sel(GLB_PKA_CLK_MCU_MUXPLL_160M);
-
-#ifdef CONFIG_BSP_SDH_SDCARD
-    PERIPHERAL_CLOCK_SDH_ENABLE();
-    uint32_t tmp_val;
-    tmp_val = BL_RD_REG(PDS_BASE, PDS_CTL5);
-    uint32_t tmp_val2 = BL_GET_REG_BITS_VAL(tmp_val, PDS_CR_PDS_GPIO_KEEP_EN);
-    tmp_val2 &= ~(1 << 0);
-    tmp_val = BL_SET_REG_BITS_VAL(tmp_val, PDS_CR_PDS_GPIO_KEEP_EN, tmp_val2);
-    BL_WR_REG(PDS_BASE, PDS_CTL5, tmp_val);
-    GLB_AHB_MCU_Software_Reset(GLB_AHB_MCU_SW_SDH);
-#endif
-#ifdef CONFIG_BSP_CSI
-    GLB_CSI_Config_MIPIPLL(2, 0x21000);
-    GLB_CSI_Power_Up_MIPIPLL();
-    GLB_Set_DSP_CLK(ENABLE, GLB_DSP_CLK_MUXPLL_160M, 1);
-#endif
-    GLB_Set_USB_CLK_From_WIFIPLL(1);
-}
-
-#ifdef CONFIG_PSRAM
-#define WB_4MB_PSRAM   (1)
-#define UHS_32MB_PSRAM (2)
-#define UHS_64MB_PSRAM (3)
-#define WB_32MB_PSRAM  (4)
-#define NONE_UHS_PSRAM (-1)
-
-int uhs_psram_init(void)
-{
-    PSRAM_UHS_Cfg_Type psramDefaultCfg = {
-        2000,
-        PSRAM_MEM_SIZE_32MB,
-        PSRAM_PAGE_SIZE_2KB,
-        PSRAM_UHS_NORMAL_TEMP,
-    };
-
-    bflb_efuse_device_info_type chip_info;
-    bflb_ef_ctrl_get_device_info(&chip_info);
-    if (chip_info.psramInfo == UHS_32MB_PSRAM) {
-        psramDefaultCfg.psramMemSize = PSRAM_MEM_SIZE_32MB;
-    } else if (chip_info.psramInfo == UHS_64MB_PSRAM) {
-        psramDefaultCfg.psramMemSize = PSRAM_MEM_SIZE_64MB;
-    } else {
-        return -1;
-    }
-
-    //init uhs PLL; Must open uhs pll first, and then initialize uhs psram
-    GLB_Config_UHS_PLL(GLB_XTAL_40M, uhsPllCfg_2000M);
-    //init uhs psram ;
-    // Psram_UHS_x16_Init(Clock_Peripheral_Clock_Get(BL_PERIPHERAL_CLOCK_PSRAMA) / 1000000);
-    Psram_UHS_x16_Init_Override(&psramDefaultCfg);
-    Tzc_Sec_PSRAMA_Access_Release();
-
-    // example: 2000Mbps typical cal values
-    uhs_phy_cal_res->rl = 39;
-    uhs_phy_cal_res->rdqs = 3;
-    uhs_phy_cal_res->rdq = 0;
-    uhs_phy_cal_res->wl = 13;
-    uhs_phy_cal_res->wdqs = 4;
-    uhs_phy_cal_res->wdq = 5;
-    uhs_phy_cal_res->ck = 9;
-    /* TODO: use uhs psram trim update */
-    set_uhs_latency_r(uhs_phy_cal_res->rl);
-    cfg_dqs_rx(uhs_phy_cal_res->rdqs);
-    cfg_dq_rx(uhs_phy_cal_res->rdq);
-    set_uhs_latency_w(uhs_phy_cal_res->wl);
-    cfg_dq_drv(uhs_phy_cal_res->wdq);
-    cfg_ck_cen_drv(uhs_phy_cal_res->wdq + 4, uhs_phy_cal_res->wdq + 1);
-    cfg_dqs_drv(uhs_phy_cal_res->wdqs);
-    // set_odt_en();
-    mr_read_back();
-    return 0;
-}
-#endif
-#endif
-
-void bl_show_log(void)
-{
-    printf("\r\n");
-    printf("  ____               __  __      _       _       _     \r\n");
-    printf(" |  _ \\             / _|/ _|    | |     | |     | |    \r\n");
-    printf(" | |_) | ___  _   _| |_| |_ __ _| | ___ | | __ _| |__  \r\n");
-    printf(" |  _ < / _ \\| | | |  _|  _/ _` | |/ _ \\| |/ _` | '_ \\ \r\n");
-    printf(" | |_) | (_) | |_| | | | || (_| | | (_) | | (_| | |_) |\r\n");
-    printf(" |____/ \\___/ \\__,_|_| |_| \\__,_|_|\\___/|_|\\__,_|_.__/ \r\n");
-    printf("\r\n");
-    printf("Build:%s,%s\r\n", __TIME__, __DATE__);
-    printf("Copyright (c) 2022 Bouffalolab team\r\n");
-}
-
-void bl_show_flashinfo(void)
-{
-    spi_flash_cfg_type flashCfg;
-    uint8_t *pFlashCfg = NULL;
-    uint32_t flashCfgLen = 0;
-    uint32_t flashJedecId = 0;
-
-    flashJedecId = bflb_flash_get_jedec_id();
-    bflb_flash_get_cfg(&pFlashCfg, &flashCfgLen);
-    arch_memcpy((void *)&flashCfg, pFlashCfg, flashCfgLen);
-    printf("=========== flash cfg ==============\r\n");
-    printf("jedec id   0x%06X\r\n", flashJedecId);
-    printf("mid            0x%02X\r\n", flashCfg.mid);
-    printf("iomode         0x%02X\r\n", flashCfg.io_mode);
-    printf("clk delay      0x%02X\r\n", flashCfg.clk_delay);
-    printf("clk invert     0x%02X\r\n", flashCfg.clk_invert);
-    printf("read reg cmd0  0x%02X\r\n", flashCfg.read_reg_cmd[0]);
-    printf("read reg cmd1  0x%02X\r\n", flashCfg.read_reg_cmd[1]);
-    printf("write reg cmd0 0x%02X\r\n", flashCfg.write_reg_cmd[0]);
-    printf("write reg cmd1 0x%02X\r\n", flashCfg.write_reg_cmd[1]);
-    printf("qe write len   0x%02X\r\n", flashCfg.qe_write_reg_len);
-    printf("cread support  0x%02X\r\n", flashCfg.c_read_support);
-    printf("cread code     0x%02X\r\n", flashCfg.c_read_mode);
-    printf("burst wrap cmd 0x%02X\r\n", flashCfg.burst_wrap_cmd);
-    printf("=====================================\r\n");
-}
-
-extern void bflb_uart_set_console(struct bflb_device_s *dev);
-
-static void console_init()
-{
-    struct bflb_device_s *gpio;
-
-    gpio = bflb_device_get_by_name("gpio");
-#if defined(CPU_M0)
-    bflb_gpio_uart_init(gpio, GPIO_PIN_14, GPIO_UART_FUNC_UART0_TX);
-    bflb_gpio_uart_init(gpio, GPIO_PIN_15, GPIO_UART_FUNC_UART0_RX);
-#elif defined(CPU_D0)
-    /* sipeed m1s dock */
-    bflb_gpio_init(gpio, GPIO_PIN_16, 21 | GPIO_ALTERNATE | GPIO_PULLUP | GPIO_SMT_EN | GPIO_DRV_1);
-    bflb_gpio_init(gpio, GPIO_PIN_17, 21 | GPIO_ALTERNATE | GPIO_PULLUP | GPIO_SMT_EN | GPIO_DRV_1);
-#endif
-    struct bflb_uart_config_s cfg;
-    cfg.baudrate = 2000000;
-    cfg.data_bits = UART_DATA_BITS_8;
-    cfg.stop_bits = UART_STOP_BITS_1;
-    cfg.parity = UART_PARITY_NONE;
-    cfg.flow_ctrl = 0;
-    cfg.tx_fifo_threshold = 7;
-    cfg.rx_fifo_threshold = 7;
-#if defined(CPU_M0)
-    uart0 = bflb_device_get_by_name("uart0");
-#elif defined(CPU_D0)
-    uart0 = bflb_device_get_by_name("uart3");
-#endif
-    bflb_uart_init(uart0, &cfg);
-    bflb_uart_set_console(uart0);
-}
-
-#if defined(CPU_M0)
-void board_init(void)
-{
-    int ret = -1;
-    uintptr_t flag;
-
-    flag = bflb_irq_save();
-
-    ret = bflb_flash_init();
-
-    GLB_Halt_CPU(GLB_CORE_ID_D0);
-
-    system_clock_init();
-    peripheral_clock_init();
-    bflb_irq_initialize();
-
-    console_init();
-
-    size_t heap_len = ((size_t)&__HeapLimit - (size_t)&__HeapBase);
-    kmem_init((void *)&__HeapBase, heap_len);
-
-    bl_show_log();
-    if (ret != 0) {
-        printf("flash init fail!!!\r\n");
-    }
-    bl_show_flashinfo();
-
-    printf("dynamic memory init success,heap size = %d Kbyte \r\n", ((size_t)&__HeapLimit - (size_t)&__HeapBase) / 1024);
-
-    printf("sig1:%08x\r\n", BL_RD_REG(GLB_BASE, GLB_UART_CFG1));
-    printf("sig2:%08x\r\n", BL_RD_REG(GLB_BASE, GLB_UART_CFG2));
-
-    log_start();
-
-#if (defined(CONFIG_LUA) || defined(CONFIG_BFLOG) || defined(CONFIG_FATFS))
-    rtc = bflb_device_get_by_name("rtc");
-#endif
-
-#ifdef CONFIG_PSRAM
-    if (uhs_psram_init() < 0) {
-        while (1) {
-        }
-    }
-#endif
-    /* set CPU D0 boot XIP address and flash address */
-    // Tzc_Sec_Set_CPU_Group(GLB_CORE_ID_D0, 1);
-    // /* D0 boot from 0x58000000 */
-    // GLB_Set_CPU_Reset_Address(GLB_CORE_ID_D0, 0x58000000);
-    // /* D0 image offset on flash is 0x100000+0x1000(header) */
-    // bflb_sf_ctrl_set_flash_image_offset(0x101000, 1, SF_CTRL_FLASH_BANK0);
-
-    bflb_irq_restore(flag);
-
-    /* we do not check header at 0x10000, just boot */
-    GLB_Release_CPU(GLB_CORE_ID_D0);
-
-    /* release d0 and then do can run */
-    BL_WR_WORD(IPC_SYNC_ADDR1, IPC_SYNC_FLAG);
-    BL_WR_WORD(IPC_SYNC_ADDR2, IPC_SYNC_FLAG);
-    L1C_DCache_Clean_By_Addr(IPC_SYNC_ADDR1, 8);
-}
-#elif defined(CPU_D0)
-void board_init(void)
-{
-    CPU_Set_MTimer_CLK(ENABLE, CPU_Get_MTimer_Source_Clock() / 1000 / 1000 - 1);
-
-    bflb_irq_initialize();
-
-    size_t heap_len = ((size_t)&__HeapLimit - (size_t)&__HeapBase);
-    kmem_init((void *)&__HeapBase, heap_len);
-
-    console_init();
-
-    bl_show_log();
-
-    printf("dynamic memory init success,heap size = %d Kbyte \r\n", ((size_t)&__HeapLimit - (size_t)&__HeapBase) / 1024);
-
-    printf("sig1:%08x\r\n", BL_RD_REG(GLB_BASE, GLB_UART_CFG1));
-    printf("sig2:%08x\r\n", BL_RD_REG(GLB_BASE, GLB_UART_CFG2));
-    printf("cgen1:%08x\r\n", getreg32(BFLB_GLB_CGEN1_BASE));
-
-    log_start();
-}
-#endif
-
-void board_uartx_gpio_init(void)
-{
-    struct bflb_device_s *gpio;
-
-    gpio = bflb_device_get_by_name("gpio");
-
-    bflb_gpio_uart_init(gpio, GPIO_PIN_4, GPIO_UART_FUNC_UART1_TX);
-    bflb_gpio_uart_init(gpio, GPIO_PIN_5, GPIO_UART_FUNC_UART1_RX);
-    bflb_gpio_uart_init(gpio, GPIO_PIN_6, GPIO_UART_FUNC_UART1_CTS);
-    bflb_gpio_uart_init(gpio, GPIO_PIN_7, GPIO_UART_FUNC_UART1_RTS);
-}
-
-void board_i2c0_gpio_init(void)
-{
-    struct bflb_device_s *gpio;
-
-    gpio = bflb_device_get_by_name("gpio");
-    /* I2C0_SDA */
-    bflb_gpio_init(gpio, GPIO_PIN_11, GPIO_FUNC_I2C0 | GPIO_ALTERNATE | GPIO_PULLUP | GPIO_SMT_EN | GPIO_DRV_1);
-    /* I2C0_SCL */
-    bflb_gpio_init(gpio, GPIO_PIN_16, GPIO_FUNC_I2C0 | GPIO_ALTERNATE | GPIO_PULLUP | GPIO_SMT_EN | GPIO_DRV_1);
-}
-
-void board_spi0_gpio_init(void)
-{
-    struct bflb_device_s *gpio;
-
-    gpio = bflb_device_get_by_name("gpio");
-    /* spi cs */
-    bflb_gpio_init(gpio, GPIO_PIN_16, GPIO_FUNC_SPI0 | GPIO_ALTERNATE | GPIO_PULLUP | GPIO_SMT_EN | GPIO_DRV_1);
-    /* spi miso */
-    bflb_gpio_init(gpio, GPIO_PIN_17, GPIO_FUNC_SPI0 | GPIO_ALTERNATE | GPIO_PULLUP | GPIO_SMT_EN | GPIO_DRV_1);
-    /* spi mosi */
-    bflb_gpio_init(gpio, GPIO_PIN_18, GPIO_FUNC_SPI0 | GPIO_ALTERNATE | GPIO_PULLUP | GPIO_SMT_EN | GPIO_DRV_1);
-    /* spi clk */
-    bflb_gpio_init(gpio, GPIO_PIN_19, GPIO_FUNC_SPI0 | GPIO_ALTERNATE | GPIO_PULLUP | GPIO_SMT_EN | GPIO_DRV_1);
-}
-
-void board_pwm_gpio_init(void)
-{
-    struct bflb_device_s *gpio;
-
-    gpio = bflb_device_get_by_name("gpio");
-    bflb_gpio_init(gpio, GPIO_PIN_24, GPIO_FUNC_PWM0 | GPIO_ALTERNATE | GPIO_PULLDOWN | GPIO_SMT_EN | GPIO_DRV_1);
-    bflb_gpio_init(gpio, GPIO_PIN_25, GPIO_FUNC_PWM0 | GPIO_ALTERNATE | GPIO_PULLUP | GPIO_SMT_EN | GPIO_DRV_1);
-    bflb_gpio_init(gpio, GPIO_PIN_26, GPIO_FUNC_PWM0 | GPIO_ALTERNATE | GPIO_PULLDOWN | GPIO_SMT_EN | GPIO_DRV_1);
-    bflb_gpio_init(gpio, GPIO_PIN_27, GPIO_FUNC_PWM0 | GPIO_ALTERNATE | GPIO_PULLUP | GPIO_SMT_EN | GPIO_DRV_1);
-    bflb_gpio_init(gpio, GPIO_PIN_28, GPIO_FUNC_PWM0 | GPIO_ALTERNATE | GPIO_PULLDOWN | GPIO_SMT_EN | GPIO_DRV_1);
-    bflb_gpio_init(gpio, GPIO_PIN_29, GPIO_FUNC_PWM0 | GPIO_ALTERNATE | GPIO_PULLUP | GPIO_SMT_EN | GPIO_DRV_1);
-    bflb_gpio_init(gpio, GPIO_PIN_30, GPIO_FUNC_PWM0 | GPIO_ALTERNATE | GPIO_PULLDOWN | GPIO_SMT_EN | GPIO_DRV_1);
-    bflb_gpio_init(gpio, GPIO_PIN_31, GPIO_FUNC_PWM0 | GPIO_ALTERNATE | GPIO_PULLUP | GPIO_SMT_EN | GPIO_DRV_1);
-}
-
-void board_adc_gpio_init(void)
-{
-    struct bflb_device_s *gpio;
-
-    gpio = bflb_device_get_by_name("gpio");
-    /* ADC_CH0 */
-    bflb_gpio_init(gpio, GPIO_PIN_17, GPIO_ANALOG | GPIO_SMT_EN | GPIO_DRV_0);
-    /* ADC_CH1 */
-    bflb_gpio_init(gpio, GPIO_PIN_5, GPIO_ANALOG | GPIO_SMT_EN | GPIO_DRV_0);
-    /* ADC_CH2 */
-    bflb_gpio_init(gpio, GPIO_PIN_4, GPIO_ANALOG | GPIO_SMT_EN | GPIO_DRV_0);
-    /* ADC_CH3 */
-    bflb_gpio_init(gpio, GPIO_PIN_11, GPIO_ANALOG | GPIO_SMT_EN | GPIO_DRV_0);
-    /* ADC_CH4 */
-    bflb_gpio_init(gpio, GPIO_PIN_6, GPIO_ANALOG | GPIO_SMT_EN | GPIO_DRV_0);
-    /* ADC_CH5 */
-    //bflb_gpio_init(gpio, GPIO_PIN_40, GPIO_ANALOG | GPIO_SMT_EN | GPIO_DRV_0);
-    /* ADC_CH6 */
-    bflb_gpio_init(gpio, GPIO_PIN_12, GPIO_ANALOG | GPIO_SMT_EN | GPIO_DRV_0);
-    /* ADC_CH7 */
-    bflb_gpio_init(gpio, GPIO_PIN_13, GPIO_ANALOG | GPIO_SMT_EN | GPIO_DRV_0);
-    /* ADC_CH8 */
-    bflb_gpio_init(gpio, GPIO_PIN_16, GPIO_ANALOG | GPIO_SMT_EN | GPIO_DRV_0);
-    /* ADC_CH9 */
-    bflb_gpio_init(gpio, GPIO_PIN_18, GPIO_ANALOG | GPIO_SMT_EN | GPIO_DRV_0);
-    /* ADC_CH10 */
-    bflb_gpio_init(gpio, GPIO_PIN_19, GPIO_ANALOG | GPIO_SMT_EN | GPIO_DRV_0);
-    /* ADC_CH11,note the FLASH pin */
-    //bflb_gpio_init(gpio, GPIO_PIN_34, GPIO_ANALOG | GPIO_SMT_EN | GPIO_DRV_0);
-}
-
-void board_dac_gpio_init(void)
-{
-    struct bflb_device_s *gpio;
-
-    gpio = bflb_device_get_by_name("gpio");
-    /* DAC_CHA */
-    bflb_gpio_init(gpio, GPIO_PIN_11, GPIO_ANALOG | GPIO_SMT_EN | GPIO_DRV_0);
-    /* DAC_CHB */
-    bflb_gpio_init(gpio, GPIO_PIN_4, GPIO_ANALOG | GPIO_SMT_EN | GPIO_DRV_0);
-}
-
-void board_ir_gpio_init(void)
-{
-    struct bflb_device_s *gpio;
-
-    gpio = bflb_device_get_by_name("gpio");
-    bflb_gpio_init(gpio, GPIO_PIN_11, GPIO_ANALOG | GPIO_SMT_EN | GPIO_DRV_0);
-    bflb_gpio_init(gpio, GPIO_PIN_17, GPIO_INPUT | GPIO_SMT_EN | GPIO_DRV_0);
-    GLB_IR_RX_GPIO_Sel(GLB_GPIO_PIN_17);
-}
-
-void board_emac_gpio_init(void)
-{
-    struct bflb_device_s *gpio;
-
-    gpio = bflb_device_get_by_name("gpio");
-    bflb_gpio_init(gpio, GPIO_PIN_24, GPIO_FUNC_EMAC | GPIO_ALTERNATE | GPIO_PULLUP | GPIO_SMT_EN | GPIO_DRV_1);
-    bflb_gpio_init(gpio, GPIO_PIN_25, GPIO_FUNC_EMAC | GPIO_ALTERNATE | GPIO_PULLUP | GPIO_SMT_EN | GPIO_DRV_1);
-    bflb_gpio_init(gpio, GPIO_PIN_26, GPIO_FUNC_EMAC | GPIO_ALTERNATE | GPIO_PULLUP | GPIO_SMT_EN | GPIO_DRV_1);
-    bflb_gpio_init(gpio, GPIO_PIN_27, GPIO_FUNC_EMAC | GPIO_ALTERNATE | GPIO_PULLUP | GPIO_SMT_EN | GPIO_DRV_1);
-    bflb_gpio_init(gpio, GPIO_PIN_28, GPIO_FUNC_EMAC | GPIO_ALTERNATE | GPIO_PULLUP | GPIO_SMT_EN | GPIO_DRV_1);
-    bflb_gpio_init(gpio, GPIO_PIN_29, GPIO_FUNC_EMAC | GPIO_ALTERNATE | GPIO_PULLUP | GPIO_SMT_EN | GPIO_DRV_1);
-    bflb_gpio_init(gpio, GPIO_PIN_30, GPIO_FUNC_EMAC | GPIO_ALTERNATE | GPIO_PULLUP | GPIO_SMT_EN | GPIO_DRV_1);
-    bflb_gpio_init(gpio, GPIO_PIN_31, GPIO_FUNC_EMAC | GPIO_ALTERNATE | GPIO_PULLUP | GPIO_SMT_EN | GPIO_DRV_1);
-    bflb_gpio_init(gpio, GPIO_PIN_32, GPIO_FUNC_EMAC | GPIO_ALTERNATE | GPIO_PULLUP | GPIO_SMT_EN | GPIO_DRV_1);
-    bflb_gpio_init(gpio, GPIO_PIN_33, GPIO_FUNC_EMAC | GPIO_ALTERNATE | GPIO_PULLUP | GPIO_SMT_EN | GPIO_DRV_1);
-
-#if defined(BL808)
-    // GLB_PER_Clock_UnGate(1<<12);
-#endif
-}
-
-void board_sdh_gpio_init(void)
-{
-    struct bflb_device_s *gpio;
-
-    gpio = bflb_device_get_by_name("gpio");
-    bflb_gpio_init(gpio, GPIO_PIN_0, GPIO_FUNC_SDH | GPIO_ALTERNATE | GPIO_PULLUP | GPIO_SMT_EN | GPIO_DRV_2);
-    bflb_gpio_init(gpio, GPIO_PIN_1, GPIO_FUNC_SDH | GPIO_ALTERNATE | GPIO_PULLUP | GPIO_SMT_EN | GPIO_DRV_2);
-    bflb_gpio_init(gpio, GPIO_PIN_2, GPIO_FUNC_SDH | GPIO_ALTERNATE | GPIO_PULLUP | GPIO_SMT_EN | GPIO_DRV_2);
-    bflb_gpio_init(gpio, GPIO_PIN_3, GPIO_FUNC_SDH | GPIO_ALTERNATE | GPIO_PULLUP | GPIO_SMT_EN | GPIO_DRV_2);
-    bflb_gpio_init(gpio, GPIO_PIN_4, GPIO_FUNC_SDH | GPIO_ALTERNATE | GPIO_PULLUP | GPIO_SMT_EN | GPIO_DRV_2);
-    bflb_gpio_init(gpio, GPIO_PIN_5, GPIO_FUNC_SDH | GPIO_ALTERNATE | GPIO_PULLUP | GPIO_SMT_EN | GPIO_DRV_2);
-}
-
-void board_dvp_gpio_init(void)
-{
-    struct bflb_device_s *gpio;
-
-    gpio = bflb_device_get_by_name("gpio");
-    /* I2C GPIO */
-    bflb_gpio_init(gpio, GPIO_PIN_22, GPIO_FUNC_I2C0 | GPIO_ALTERNATE | GPIO_PULLUP | GPIO_SMT_EN | GPIO_DRV_1);
-    bflb_gpio_init(gpio, GPIO_PIN_23, GPIO_FUNC_I2C0 | GPIO_ALTERNATE | GPIO_PULLUP | GPIO_SMT_EN | GPIO_DRV_1);
-
-    /* Power down GPIO */
-    bflb_gpio_init(gpio, GPIO_PIN_21, GPIO_OUTPUT | GPIO_PULLUP | GPIO_SMT_EN | GPIO_DRV_1);
-    bflb_gpio_reset(gpio, GPIO_PIN_21);
-
-    /* Reset GPIO */
-    bflb_gpio_init(gpio, GPIO_PIN_20, GPIO_OUTPUT | GPIO_PULLUP | GPIO_SMT_EN | GPIO_DRV_1);
-    bflb_gpio_set(gpio, GPIO_PIN_20);
-
-    /* MCLK GPIO */
-    bflb_gpio_init(gpio, GPIO_PIN_33, GPIO_FUNC_CLKOUT | GPIO_ALTERNATE | GPIO_PULLUP | GPIO_SMT_EN | GPIO_DRV_1);
-
-    /* DVP GPIO */
-    bflb_gpio_init(gpio, GPIO_PIN_16, GPIO_FUNC_CAM | GPIO_ALTERNATE | GPIO_PULLUP | GPIO_SMT_EN | GPIO_DRV_1);
-    bflb_gpio_init(gpio, GPIO_PIN_17, GPIO_FUNC_CAM | GPIO_ALTERNATE | GPIO_PULLUP | GPIO_SMT_EN | GPIO_DRV_1);
-    bflb_gpio_init(gpio, GPIO_PIN_24, GPIO_FUNC_CAM | GPIO_ALTERNATE | GPIO_PULLUP | GPIO_SMT_EN | GPIO_DRV_1);
-    bflb_gpio_init(gpio, GPIO_PIN_25, GPIO_FUNC_CAM | GPIO_ALTERNATE | GPIO_PULLUP | GPIO_SMT_EN | GPIO_DRV_1);
-    bflb_gpio_init(gpio, GPIO_PIN_26, GPIO_FUNC_CAM | GPIO_ALTERNATE | GPIO_PULLUP | GPIO_SMT_EN | GPIO_DRV_1);
-    bflb_gpio_init(gpio, GPIO_PIN_27, GPIO_FUNC_CAM | GPIO_ALTERNATE | GPIO_PULLUP | GPIO_SMT_EN | GPIO_DRV_1);
-    bflb_gpio_init(gpio, GPIO_PIN_28, GPIO_FUNC_CAM | GPIO_ALTERNATE | GPIO_PULLUP | GPIO_SMT_EN | GPIO_DRV_1);
-    bflb_gpio_init(gpio, GPIO_PIN_29, GPIO_FUNC_CAM | GPIO_ALTERNATE | GPIO_PULLUP | GPIO_SMT_EN | GPIO_DRV_1);
-    bflb_gpio_init(gpio, GPIO_PIN_30, GPIO_FUNC_CAM | GPIO_ALTERNATE | GPIO_PULLUP | GPIO_SMT_EN | GPIO_DRV_1);
-    bflb_gpio_init(gpio, GPIO_PIN_31, GPIO_FUNC_CAM | GPIO_ALTERNATE | GPIO_PULLUP | GPIO_SMT_EN | GPIO_DRV_1);
-    bflb_gpio_init(gpio, GPIO_PIN_32, GPIO_FUNC_CAM | GPIO_ALTERNATE | GPIO_PULLUP | GPIO_SMT_EN | GPIO_DRV_1);
-}
-
-void board_csi_gpio_init(void)
-{
-    struct bflb_device_s *gpio;
-
-    gpio = bflb_device_get_by_name("gpio");
-
-    GLB_Set_Ldo15cis_Vout(GLB_LDO15CIS_LEVEL_1P20V);
-#if 1 /* sipeed m1s dock */
-    /* I2C GPIO */
-    bflb_gpio_init(gpio, GPIO_PIN_6, GPIO_FUNC_I2C0 | GPIO_ALTERNATE | GPIO_PULLUP | GPIO_SMT_EN | GPIO_DRV_1);
-    bflb_gpio_init(gpio, GPIO_PIN_7, GPIO_FUNC_I2C0 | GPIO_ALTERNATE | GPIO_PULLUP | GPIO_SMT_EN | GPIO_DRV_1);
-
-    /* MCLK GPIO */
-    bflb_gpio_init(gpio, GPIO_PIN_33, GPIO_FUNC_CLKOUT | GPIO_ALTERNATE | GPIO_PULLUP | GPIO_SMT_EN | GPIO_DRV_1);
-
-    /* Power down GPIO */
-    bflb_gpio_init(gpio, GPIO_PIN_40, GPIO_OUTPUT | GPIO_PULLUP | GPIO_SMT_EN | GPIO_DRV_1);
-    bflb_gpio_reset(gpio, GPIO_PIN_40);
-
-    /* Reset GPIO */
-    bflb_gpio_init(gpio, GPIO_PIN_41, GPIO_OUTPUT | GPIO_PULLUP | GPIO_SMT_EN | GPIO_DRV_1);
-    bflb_mtimer_delay_us(20);
-    bflb_gpio_set(gpio, GPIO_PIN_41);
-#else
-    /* I2C GPIO */
-    bflb_gpio_init(gpio, GPIO_PIN_21, GPIO_FUNC_I2C0 | GPIO_ALTERNATE | GPIO_PULLUP | GPIO_SMT_EN | GPIO_DRV_1);
-    bflb_gpio_init(gpio, GPIO_PIN_22, GPIO_FUNC_I2C0 | GPIO_ALTERNATE | GPIO_PULLUP | GPIO_SMT_EN | GPIO_DRV_1);
-
-    /* MCLK GPIO */
-    bflb_gpio_init(gpio, GPIO_PIN_18, GPIO_FUNC_CLKOUT | GPIO_ALTERNATE | GPIO_PULLUP | GPIO_SMT_EN | GPIO_DRV_1);
-
-    /* Power down GPIO */
-    bflb_gpio_init(gpio, GPIO_PIN_6, GPIO_OUTPUT | GPIO_PULLUP | GPIO_SMT_EN | GPIO_DRV_1);
-    bflb_gpio_reset(gpio, GPIO_PIN_6);
-
-    /* Reset GPIO */
-    bflb_gpio_init(gpio, GPIO_PIN_23, GPIO_OUTPUT | GPIO_PULLUP | GPIO_SMT_EN | GPIO_DRV_1);
-    bflb_mtimer_delay_us(20);
-    bflb_gpio_set(gpio, GPIO_PIN_23);
-#endif
-}
-
-void board_iso11898_gpio_init()
-{
-    // struct bflb_device_s *gpio;
-
-    // gpio = bflb_device_get_by_name("gpio");
-}
-
-#ifdef CONFIG_BFLOG
-__attribute__((weak)) uint64_t bflog_clock(void)
-{
-    return bflb_mtimer_get_time_us();
-}
-
-__attribute__((weak)) uint32_t bflog_time(void)
-{
-    return BFLB_RTC_TIME2SEC(bflb_rtc_get_time(rtc));
-}
-
-__attribute__((weak)) char *bflog_thread(void)
-{
-    return "";
-}
-#endif
-
-#ifdef CONFIG_LUA
-__attribute__((weak)) clock_t luaport_clock(void)
-{
-    return (clock_t)bflb_mtimer_get_time_us();
-}
-
-__attribute__((weak)) time_t luaport_time(time_t *seconds)
-{
-    time_t t = (time_t)BFLB_RTC_TIME2SEC(bflb_rtc_get_time(rtc));
-    if (seconds != NULL) {
-        *seconds = t;
-    }
-
-    return t;
-}
-#endif
-
-#ifdef CONFIG_FATFS
-#include "bflb_timestamp.h"
-__attribute__((weak)) uint32_t get_fattime(void)
-{
-    bflb_timestamp_t tm;
-
-    bflb_timestamp_utc2time(BFLB_RTC_TIME2SEC(bflb_rtc_get_time(rtc)), &tm);
-
-    return ((uint32_t)(tm.year - 1980) << 25) /* Year 2015 */
-           | ((uint32_t)tm.mon << 21)         /* Month 1 */
-           | ((uint32_t)tm.mday << 16)        /* Mday 1 */
-           | ((uint32_t)tm.hour << 11)        /* Hour 0 */
-           | ((uint32_t)tm.min << 5)          /* Min 0 */
-           | ((uint32_t)tm.sec >> 1);         /* Sec 0 */
-}
-#endif
