+++ cmake/SDKconfig
@@ -1,53 +1,65 @@
 
 mainmenu "Pine64 OX64 SDK configuration"
+config BFLB_CHIP
+	string
+	default "$(BFLB_CHIP)"
+
+config BFLB_BOARD
+	string
+	default "$(BFLB_BOARD)"
+
+config BFLB_CPU_ID
+	string
+	default "$(BFLB_CPU_ID)"
+
+comment "Configuration for $(BFLB_CHIP) $(BFLB_CPU_ID)"
 
 menu "Application"
 	osource "Kconfig"
 endmenu
 
 menu "Board Support Configuration"
-	menu "$(BFLB_BOARD) Board Config"
-		source "$(BFLB_BOARD_DIR)/$(BFLB_BOARD)/Kconfig"
-	endmenu
-
+	osource "$(BFLB_BOARD_DIR)/$(BFLB_BOARD)/Kconfig"
+	osource "$(BFLB_BOARD_DIR)/common/$(BFLB_CHIP)/Kconfig"
+	
 	config PSRAM
 		bool "Enable PSRAM"
 		default y
 
 	config ETHERNET
-		bool "Enable Ethernet"
+		bool "Enable Ethernet SDK Driver"
 		default n
 	
 	config BSP_SDH_SDCARD
-		bool "Enable SD SDHCI Driver Support"
-		default n
-
-	config BSP_SPI_SDCARD
-		bool "Enable SD SPI Driver Support"
+		bool "Enable SD SDHCI SDK Driver"
+		depends on BOARD_HAS_SDCARD
+		select PINMUX_ENABLE_SDH
 		default n
 	
 	config BSP_LCD
-		bool "Enable LCD Driver Support"
+		bool "Enable LCD SDK Driver"
+		default n
+
+	config BSP_TOUCH
+		bool "Enable Touch SDK Driver"
 		default n
 
 	config BSP_IMAGE_SENSOR
-		bool "Enable Image Sensor Driver Support"
+		bool "Enable Image Sensor SDK"
 		default n
+	
+	config BSP_CSI
+		bool "Enable CSI SDK Driver"
+		default n
+		depends on BSP_IMAGE_SENSOR
+
 endmenu
 
 menu "CPU Configuration"
-	config BFLB_CHIP
-		string
-		default "$(BFLB_CHIP)"
-	config BLFB_CPU_ID
-		string
-		default "$(BFLB_CPU_ID)"
-	
 	config ROMAPI
 		bool "Enable ROM API"
 		default y
 endmenu
-
 
 menu "SDK configuration"
 	menu "Logging Configuration"
